Refactor react-query hooks
