+++ src/hooks/chat.ts
@@ -1,56 +1,25 @@
 import { useEffect, useRef, useState } from 'react';
-import { useRecoilState, useRecoilValue, useSetRecoilState } from 'recoil';
-import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
+import { useRecoilValue } from 'recoil';
+import { useQueryClient } from '@tanstack/react-query';
 import { io } from 'socket.io-client';
+import { toast } from 'react-toastify';
 import { accessTokenState } from '@/states/user/auth';
-import { RoomInfo, roomInfoState } from '@/states/roomInfoState';
-import { Friend } from '@/hooks/friends';
-import { useFetcher } from '@/hooks/fetcher';
+import { Friend } from '@/hooks/query/friends';
+import { Message } from '@/hooks/query/chat';
 import {
   useDismiss,
   useFloating,
   useInteractions,
 } from '@floating-ui/react-dom-interactions';
-import { toast } from 'react-toastify';
-
-export type RoomListType = 'all' | 'joined';
 
 interface SubscribeData {
   sender: Friend;
   payload: string;
 }
 
-interface Message {
-  room_msg_id: number;
-  room_id: string;
-  sender: Friend;
-  payload: string;
-  created: Date;
-}
-
-export function useChatMessages(roomId: string) {
-  const fetcher = useFetcher();
-  const data = useQuery<Message[]>({
-    queryKey: ['chat/room/messages', roomId],
-    queryFn: async () => {
-      const res = await fetcher(`/chat/room/${roomId}/messages`);
-
-      if (res.ok) {
-        const data = await res.json();
-        return data.messages;
-      }
-      return [];
-    },
-    refetchOnWindowFocus: false,
-  });
-
-  return data;
-}
-
 export function useChat(roomId: string) {
   const queryClient = useQueryClient();
   const socketRef = useSocketRef(`ws://${import.meta.env.VITE_BASE_URL}:9999`);
-  const { data: messages } = useChatMessages(roomId);
 
   useEffect(() => {
     const socket = socketRef.current;
@@ -91,174 +60,7 @@ export function useChat(roomId: string) {
     };
   }, [roomId, queryClient, socketRef]);
 
-  return { messages, socket: socketRef.current };
-}
-
-export function useRoomList(type: RoomListType) {
-  const queryClient = useQueryClient();
-  const fetcher = useFetcher();
-  const result = useQuery({
-    queryKey: ['chat/rooms', type],
-    queryFn: async (): Promise<RoomInfo[]> => {
-      const res = await fetcher(`/chat/rooms/${type}`);
-      if (res.ok) {
-        const data = await res.json();
-        return data.rooms;
-      }
-      return [];
-    },
-    select: (roomList) => {
-      if (type === 'all') {
-        const joinedRoomList = queryClient.getQueryData<RoomInfo[]>([
-          'chat/rooms',
-          'joined',
-        ]);
-
-        if (joinedRoomList) {
-          return roomList.filter(
-            (all) =>
-              !joinedRoomList.find((joined) => joined.room_id === all.room_id),
-          );
-        }
-      }
-      return roomList;
-    },
-  });
-  return result;
-}
-
-export function useJoinRoom() {
-  const fetcher = useFetcher();
-  const setRoomInfo = useSetRecoilState(roomInfoState);
-  const mutation = useMutation({
-    mutationFn: async ({
-      room,
-      password = '',
-    }: {
-      room: RoomInfo;
-      password?: string;
-    }) => {
-      return fetcher(`/chat/room/${room.room_id}/join`, {
-        method: 'POST',
-        body: JSON.stringify({ room_password: password }),
-      });
-    },
-    onSuccess: (data, info) => {
-      if (data.ok) {
-        setRoomInfo(info.room ?? null);
-      } else {
-        throw data;
-      }
-    },
-    onError: () => {
-      toast.error('Fail to join this room');
-    },
-  });
-  return mutation;
-}
-
-export function useExitRoom() {
-  const fetcher = useFetcher();
-  const setRoomInfo = useSetRecoilState(roomInfoState);
-  const mutation = useMutation({
-    mutationFn: async (room_id: string) => {
-      return fetcher(`/chat/room/${room_id}/exit`, {
-        method: 'POST',
-      });
-    },
-    onSuccess: (data) => {
-      if (data.ok) {
-        setRoomInfo(null);
-      } else {
-        throw data;
-      }
-    },
-    onError: () => {
-      toast.error('Fail to exit this room');
-    },
-  });
-  return mutation;
-}
-
-export function useDeleteRoom() {
-  const fetcher = useFetcher();
-  const setRoomInfo = useSetRecoilState(roomInfoState);
-  const mutation = useMutation({
-    mutationFn: async (room_id: string) => {
-      return fetcher(`/chat/room/${room_id}`, {
-        method: 'DELETE',
-      });
-    },
-    onSuccess: (data) => {
-      if (data.ok) {
-        setRoomInfo(null);
-      } else {
-        throw data;
-      }
-    },
-    onError: () => {
-      toast.error('Fail to delete this room');
-    },
-  });
-  return mutation;
-}
-
-export function useEditRoomAccess(room_id: string) {
-  const fetcher = useFetcher();
-  const [roomInfo, setRoomInfo] = useRecoilState(roomInfoState);
-  const mutation = useMutation({
-    mutationFn: (room_access: RoomInfo['room_access']) => {
-      return fetcher(`/chat/room/${room_id}/access`, {
-        method: 'PUT',
-        body: JSON.stringify({ room_access }),
-      });
-    },
-    onSuccess: (res, room_access) => {
-      if (!res.ok) throw res;
-      setRoomInfo((currRoomInfo) =>
-        currRoomInfo
-          ? ({
-              ...roomInfo,
-              room_access,
-            } as RoomInfo)
-          : currRoomInfo,
-      );
-      toast.success('success');
-    },
-    onError: () => {
-      toast.error('failed');
-    },
-  });
-  return mutation;
-}
-
-export function useEditRoomPassword(room_id: string) {
-  const fetcher = useFetcher();
-  const [roomInfo, setRoomInfo] = useRecoilState(roomInfoState);
-  const mutation = useMutation({
-    mutationFn: (room_password: string) => {
-      return fetcher(`/chat/room/${room_id}/password`, {
-        method: 'PUT',
-        body: JSON.stringify({ room_password }),
-      });
-    },
-    onSuccess: (res, room_password) => {
-      if (!res.ok) throw res;
-      setRoomInfo((currRoomInfo) =>
-        currRoomInfo
-          ? ({
-              ...roomInfo,
-              room_password,
-            } as RoomInfo)
-          : currRoomInfo,
-      );
-      toast.success('success');
-    },
-    onError: () => {
-      toast.error('failed');
-    },
-  });
-  return mutation;
+  return { socket: socketRef.current };
 }
 
 export function usePasswordForm() {
