+++ src/hooks/profileModal.ts
@@ -6,9 +6,6 @@ import {
   useFloating,
   useInteractions,
 } from '@floating-ui/react-dom-interactions';
-import { useQuery } from '@tanstack/react-query';
-import { useFetcher } from './fetcher';
-import { UserInfo } from '@/states/user/auth';
 
 export function useProfileModal() {
   const [profileUser, setProfileUser] = useRecoilState(profileUserState);
@@ -36,20 +33,4 @@ export function useProfileModal() {
   }, [open, setProfileUser]);
 
   return { profileUser, open, setOpen, context, floating, getFloatingProps };
-}
-
-export function useUserInfo(nickname: string) {
-  const fetcher = useFetcher();
-  const data = useQuery<UserInfo>({
-    queryKey: ['user', nickname],
-    queryFn: async () => {
-      const res = await fetcher(`/user/${nickname}`);
-
-      if (res.ok) return res.json();
-      throw res;
-    },
-    retry: false,
-  });
-
-  return data;
 }
