+++ src/organism/AddFriendForm.tsx
@@ -1,36 +1,5 @@
 import { useState } from 'react';
-import { useFetcher } from '@/hooks/fetcher';
-import { useMutation, useQueryClient } from '@tanstack/react-query';
-import { toast } from 'react-toastify';
-
-function useAddFriend() {
-  const queryClient = useQueryClient();
-  const fetcher = useFetcher();
-  const mutation = useMutation({
-    mutationFn: async (nickname: string) => {
-      const res = await fetcher(`/friend/request/${nickname}`, {
-        method: 'POST',
-      });
-      if (res.ok) return res;
-      throw res;
-    },
-    onSuccess: (_, nickname) => {
-      queryClient.invalidateQueries({ queryKey: ['friend/request', 'sent'] });
-      toast.success(`sent friend request to ${nickname}`);
-    },
-    onError: (error, nickname) => {
-      let errorMsg = `friend request to ${nickname} failed`;
-
-      if (error instanceof Response) {
-        if (error.status === 404) {
-          errorMsg = `could not found ${nickname}`;
-        }
-      }
-      toast.error(errorMsg);
-    },
-  });
-  return mutation;
-}
+import { useAddFriend } from '@/hooks/mutation/friends';
 
 function AddFriendForm() {
   const [query, setQuery] = useState<string>('');
